# docs/swagger.yaml
basePath: /
definitions:
  controller.signInInput:
    properties:
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
  db.Card:
    properties:
      balance:
        type: number
      card_number:
        type: string
      id:
        type: integer
      is_blocked:
        type: boolean
      key_id:
        type: integer
      owner_name:
        type: string
    type: object
...
  db.UpdateCard:
    properties:
      balance:
        type: number
      card_number:
        type: string
      is_blocked:
        type: boolean
      key_id:
        type: integer
      owner_name:
        type: string
    type: object
...
  db.User:
    properties:
      id:
        type: integer
      is_admin:
        type: integer
      login:
        type: string
      password:
        type: string
    required:
    - login
    - password
    type: object
host: localhost:8888
info:
  contact: {}
  description: RESTful API server for MIFARE transactions authorization
  title: MIFARE API
  version: "0.1"
paths:
  /auth/sign-in:
    post:
      consumes:
      - application/json
      description: Ввод логина и пароля для получения токена доступа
      parameters:
      - description: Логин и пароль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/controller.signInInput'
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает JWT токен
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка авторизации
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Авторизация пользователя
      tags:
      - auth
  /auth/sign-up:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя в системе
      parameters:
      - description: Данные пользователя (login, password, is_admin)
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/db.User'
      produces:
      - application/json
      responses:
        "200":
          description: Возвращает ID созданного пользователя
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - auth
  /cards:
    get:
      consumes:
      - application/json
      description: Возвращает список всех карт
      produces:
      - application/json
      responses:
        "200":
          description: Список карт
          schema:
            items:
              $ref: '#/definitions/db.Card'
            type: array
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение всех карт
      tags:
      - cards
    post:
      consumes:
      - application/json
      description: Добавляет новую карту в систему
      parameters:
      - description: Данные карты
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/db.Card'
      produces:
      - application/json
      responses:
        "200":
          description: ID созданной карты
          schema:
            additionalProperties:
              type: integer
            type: object
        "400":
          description: Неверные параметры
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Создание карты
      tags:
      - cards
  /cards/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет карту по ID
      parameters:
      - description: ID карты
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Успешно удалено
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удаление карты
      tags:
      - cards
    get:
      consumes:
      - application/json
      description: Возвращает данные карты по идентификатору
      parameters:
      - description: ID карты
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Данные карты
          schema:
            $ref: '#/definitions/db.Card'
        "400":
          description: Неверный ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Получение карты по ID
      tags:
      - cards
    put:
      consumes:
      - application/json
      description: Обновляет данные существующей карты
      parameters:
      - description: ID карты
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/db.UpdateCard'
      produces:
      - application/json
      responses:
        "204":
          description: Успешно обновлено
        "400":
          description: Неверные параметры или ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Нет доступа
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Обновление карты
      tags:
      - cards
...
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"